# TaskFlow.WebApi/DTOs/TaskUpdateDto.cs
using TaskFlow.Domain.Enums;

namespace TaskFlow.Domain.DTOs;

public class TaskUpdateDto
{
    public string? Title { get; set; }
    public TaskItemStatus Status { get; set; }
    public DateTime? DueDate { get; set; }
}