# swagger.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "TaskFlow API",
      "description": "API de gestion de projets et tâches",
      "version": "v1"
    },
    "paths": {
      "/api/projects": {
        "get": {
          "tags": [
            "Projects"
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "Projects"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreateDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreateDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/projects/{id}": {
        "get": {
          "tags": [
            "Projects"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "Projects"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdateDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdateDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Projects"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/tasks": {
        "get": {
          "tags": [
            "Tasks"
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "Tasks"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/tasks/{id}": {
        "get": {
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Tasks"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/users/register": {
        "post": {
          "tags": [
            "Users"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/users/login": {
        "post": {
          "tags": [
            "Users"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "LoginDto": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "nullable": true
            },
            "password": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ProjectCreateDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ProjectUpdateDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "RegisterDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "nullable": true
            },
            "email": {
              "type": "string",
              "nullable": true
            },
            "password": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "TaskCreateDto": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "$ref": "#/components/schemas/TaskItemStatus"
            },
            "projectId": {
              "type": "integer",
              "format": "int32"
            },
            "dueDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "commentaires": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "TaskItemStatus": {
          "enum": [
            "Afaire",
            "Encours",
            "Termine"
          ],
          "type": "string"
        },
        "TaskUpdateDto": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "$ref": "#/components/schemas/TaskItemStatus"
            },
            "dueDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      },
      "securitySchemes": {
        "Bearer": {
          "type": "apiKey",
          "description": "JWT Authorization header using the Bearer scheme. Example: Bearer {token}",
          "name": "Authorization",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "Bearer": [ ]
      }
    ]
  }